Ticket 1101, phtml|php5|php4|php|cgi|pl|exe|com|bat|src| will be hardcoded additional you can set more extension in settings extended in field rename_file_types rename_file_types now is a blacklist and will no longer be renamed to .txt
update admintools modules to work with SecureForm Patch from NorHei

+++ branches/2.8.x/wb/admin/media/upload.php
@@ -66,29 +66,38 @@ if ($get_result->numRows()>0) {
 	$file_extension_string=$fetch_result['value'];
 }
 $file_extensions=explode(",",$file_extension_string);
-
+// get from settings and add to forbidden list
+$rename_file_types  = str_replace(',','|',RENAME_FILES_ON_UPLOAD);
+// hardcodet forbidden filetypes
+$forbidden_file_types = 'phtml|php5|php4|php|cgi|pl|exe|com|bat|src|'.$rename_file_types;
 // Loop through the files
 $good_uploads = 0;
 for($count = 1; $count <= 10; $count++) {
 	// If file was upload to tmp
 	if(isset($_FILES["file$count"]['name'])) {
 		// Remove bad characters
-		$filename = media_filename($_FILES["file$count"]['name']);
+		$filename = trim(media_filename($_FILES["file$count"]['name']),'.') ;
 		// Check if there is still a filename left
-		if($filename != '') {
-			// Check for potentially malicious files and append 'txt' to their name
+		// if($filename != '') {
+		$info = pathinfo($filename);
+		$ext = isset($info['extension']) ? $info['extension'] : '';
+
+		if ( ($filename != '') && !preg_match("/\." . $forbidden_file_types . "$/i", $ext) ) {
+/*
+		// Check for potentially malicious files and append 'txt' to their name
 			foreach($file_extensions as $file_ext) {
 				$file_ext_len=strlen($file_ext);
 				if (substr($filename,-$file_ext_len)==$file_ext) {
 					$filename.='.txt';
 				}
-			}		
+			}
+*/
 			// Move to relative path (in media folder)
-			if(file_exists($relative.$filename) AND $overwrite == true) {			
+			if(file_exists($relative.$filename) AND $overwrite == true) {
 				if(move_uploaded_file($_FILES["file$count"]['tmp_name'], $relative.$filename)) {
 					$good_uploads++;
 					// Chmod the uploaded file
-					change_mode($relative.$filename, 'file');
+					change_mode($relative.$filename);
 				}
 			} elseif(!file_exists($relative.$filename)) {
 				if(move_uploaded_file($_FILES["file$count"]['tmp_name'], $relative.$filename)) {
@@ -97,7 +106,7 @@ for($count = 1; $count <= 10; $count++) {
 					change_mode($relative.$filename);
 				}
 			}
-			
+
 			if(file_exists($relative.$filename)) {
 				if ($pathsettings[$resizepath]['width'] || $pathsettings[$resizepath]['height'] ) {
 					$rimg=new RESIZEIMAGE($relative.$filename);
@@ -105,7 +114,7 @@ for($count = 1; $count <= 10; $count++) {
 					$rimg->close();
 				}
 			}
-				
+
 			// store file name of first file for possible unzip action
 			if ($count == 1) {
 				$filename1 = $relative . $filename;
@@ -113,34 +122,55 @@ for($count = 1; $count <= 10; $count++) {
 		}
 	}
 }
+/*
+ * Callback function to skip files in black-list
+ */
+function pclzipCheckValidFile($p_event, &$p_header)
+{
+                         //  return 1;
+	$rename_file_types  = str_replace(',','|',RENAME_FILES_ON_UPLOAD);
+	// hardcodet forbidden filetypes
+	$forbidden_file_types = 'phtml|php5|php4|php|cgi|pl|exe|com|bat|src|'.$rename_file_types;
+	$info = pathinfo($p_header['filename']);
+                         $ext = isset($info['extension']) ? $info['extension'] : '';
+                         $dots = (substr($info['basename'], 0, 1) == '.') || (substr($info['basename'], -1, 1) == '.');
+	if( !preg_match('/'.$forbidden_file_types.'$/i', $ext) && $dots != '.' )
+	{	// ----- allowed file types are extracted
+	  return 1;
+	}else
+	{	// ----- all other files are skiped
+	  return 0;
+	}
+}
+/* ********************************* */
 
 // If the user chose to unzip the first file, unzip into the current folder
 if (isset($_POST['unzip']) && isset($filename1) && file_exists($filename1) ) {
 	$archive = new PclZip($filename1);
-	$list = $archive->extract(PCLZIP_OPT_PATH, $relative);
+
+	$list = $archive->extract(PCLZIP_OPT_PATH, $relative,PCLZIP_CB_PRE_EXTRACT, 'pclzipCheckValidFile');
+
 	if($list == 0) {
 		// error while trying to extract the archive (most likely wrong format)
 		$admin->print_error('UNABLE TO UNZIP FILE' . $archive -> errorInfo(true));
 	}
-	
+
+	$sum_dirs = 0;
+	$sum_files = 0;
+
 	// rename executable files!
-	foreach ($list as $val) {
-		$fn = $val['filename'];
-		$fnp = pathinfo($fn);
-		if (isset($fnp['extension'])) {
-			$fext = $fnp['extension'];
-			if (in_array($fext, $file_extensions)) {
-				rename($fn, $fn.".txt");
-			}
+	foreach ($list as $key => $val) {
+	    if( ($val['folder'] ) && change_mode($val['filename']) ) {
+		   $sum_dirs++;
+		} elseif( is_writable($val['filename']) && ($val['status'] == 'ok') && change_mode($val['filename']) )  {
+			$sum_files++;
 		}
 	}
+	if (isset($_POST['delzip'])) { unlink($filename1); }
 }
-
+unset($list);
 if($good_uploads == 1) {
-	$admin->print_success($good_uploads.' '.$MESSAGE['MEDIA']['SINGLE_UPLOADED'] );
-	if (isset($_POST['delzip'])) {
-		unlink($filename1);
-	}
+	$admin->print_success($sum_files.' '.$MESSAGE['MEDIA']['SINGLE_UPLOADED'] );
 } else {
 	$admin->print_success($good_uploads.' '.$MESSAGE['MEDIA']['UPLOADED'] );
 }
